+++ pluginlvox/tools/lvox_distancevisitor.h
@@ -10,7 +10,7 @@ public:
 
     LVOX_DistanceVisitor(CT_Grid3D<double> *grid);
 
-    virtual void visit(const int &index, const CT_Beam *beam);
+    virtual void visit(const size_t &index, const CT_Beam *beam);
 
 private:
     CT_Grid3D<double>*  _grid;
